# openapi3.yaml
openapi: 3.0.3
info:
  title: Cogged
  description: Cogged REST API
  version: 0.2.0
servers:
  - url: http://localhost:8090/
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      description: login using a username and password and get an auth token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
          description: ''
...
components:
  securitySchemes:
#    bearerAuth:
#      type: http
#      scheme: bearer
    bearerAuth:
      type: apiKey
      name: Authorization
      in: header      
  schemas:
...
    LoginRequest:
      nullable: false
      properties:
        password:
          type: string
          example: 'Ex4mPl3_P@55w0rd'
        username:
          type: string
          example: 'exampleuser@exampleorg.dev'
      type: object
...
    TokenResponse:
      nullable: false
      properties:
        exp:
          description: expiry time in seconds for auth token
          type: integer
          example: 600
        token:
          description: 'auth token that should be sent with subsequent requests to
            the Cogged backend in the Authorization header'
          type: string
          example: 'MHgzNC5zeXMuMTcwNTExMDAwMg.AeikLCFQtA5UfewdlN8DvakO8UvY_NibaJaPrcnIMmQ'
      type: object
...